+++ R/globals.R
@@ -12,7 +12,8 @@ utils::globalVariables(c("qol_ID", "qol_from", "qol_to", "from", "to", "delta",
                          "mean", "sd", "min", "max", "sum_wgt", ".temp_key", ".temp_weight",
                          "var_pct_row", ".pseudo_preserve", "BY", "by_vars", "VALUE",
                          "first_person", "age", "age_factor", "income", "income_factor",
-                         "expenses", "expenses_factor", "file", "result", "TYPE_ORIG"))
+                         "expenses", "expenses_factor", "file", "result", "TYPE_ORIG",
+                         "total", "value"))
 
 ###############################################################################
 # Start up