+++ NEWS.md
@@ -1,11 +1,17 @@
 # qol 1.3.2 - DEVELOPMENT
 
+### New functions
+
+* `code_statistics()`: Reads in a folder or entire folder structure, grabs all R script files, scans the contents for different patterns and outputs a small report. (04.06.2026)
+
 ### New functionality
 
 * `running_number()`: With the new `sort` parameter a vektor of variable names can be passed as `by`, which then is automatically sorted before generating the running number. There is also a new `group_nr` parameter which creates a running number for each group as a whole instead of a running number within the group. (26.05.2026)
 * `concat()`: Can now concatenate variables inside a data frame with a separator. (03.06.2026)
 * `recode_multi()`: New `convert` parameter converts recoded variables to numeric or character depending on the input format instead of leaving them as factors. (03.06.2026)
 * `any_table()`, `excel_output_style()`: Can now apply a `background_color` to all cells which aren't covered by any other background color parameter. (03.06.2026)
+* `frequencies()`: The `number format` option `sum_decimals` within the `excel_output_style()` function can now be used to set the decimal places in text outputs as well. (04.06.2026)
+* `libname()`: With the new `recursive` parameter files can now also be retreived from subfolders. (04.06.2026)
 
 ### Fixed
 
@@ -13,6 +19,7 @@
 * `print_step()`: Doesn't print empty lines anymore, when surrounding the function with `suppressMessages()`. (01.06.2026)
 * `sort_plus()`: Doesn't crash anymore when trying to preserve factor variables. (03.06.2026)
 * `export_data()`: When exporting a csv file characters like ä, ö, ü are displayed correctly, when opening the file in Excel. (03.06.2026)
+* `libname()`: When using the `extensions` parameter then now only files with these extensions are printed out to the console. Otherwise a `NOTE` is displayed stating that no file has been found. (04.06.2026)
 
 ### Additionally